✨ feat [#92] badge, profile 중복함수 -> util 함수 생성
