+++ src/components/profile/profile.tsx
@@ -1,13 +1,11 @@
 import styles from "./profile.module.css";
-import {
-  ProfileColorsArray,
-  RemainProfileColor,
-} from "@/constants/profile-random-color";
+import { ProfileColorsArray, RemainProfileColor } from "@/constants/profile-random-color";
 import { ProfileSize } from "@/components/profile/profile-size";
 import { ProfileType } from "@/components/profile/profile-type";
 import typographyStyles from "@/components/typography/typography.module.css";
 import { classnames } from "@/utils/classnames";
 import { useMemo } from "react";
+import { getColorIndex, getVisualLength } from "@/utils/string-converter";
 
 interface ProfileProps {
   size?: ProfileSize;
@@ -28,38 +26,21 @@ export default function Profile({
     if (colorIndex !== undefined) {
       return colorIndex % ProfileColorsArray.length;
     }
-
-    const hash = name.split("").reduce((acc, char, index) => {
-      return acc + char.charCodeAt(0) * (index + 1);
-    }, 0);
-
-    return hash % ProfileColorsArray.length;
+    return getColorIndex(name, ProfileColorsArray.length);
   }, [colorIndex, name]);
 
   const color = isRemain
     ? RemainProfileColor.backgroundColor
     : ProfileColorsArray[calculateColorIndex];
 
-  const getLength = (str: string) => {
-    return str.split("").reduce((acc, char) => {
-      const code = char.charCodeAt(0);
-      if (code >= 0xac00 && code <= 0xd7af) {
-        return acc + 2;
-      }
-      return acc + 1;
-    }, 0);
-  };
-
-  const visualLength = getLength(name);
-
-  const spanClasses =
-    visualLength >= 5 ? styles.alignStart : styles.alignCenter;
+  const visualLength = getVisualLength(name);
+  const spanClasses = visualLength >= 5 ? styles.alignStart : styles.alignCenter;
 
   const profileClasses = classnames(
     styles.profile,
     type === ProfileType.NavigationBar ? styles.naviType : "",
     styles[size],
-    typographyStyles.xsSemiBold
+    typographyStyles.xsSemiBold,
   );
 
   return (
